@@ -12,7 +12,7 @@ plugins {
 
 group = "team.msg"
 version = "0.0.1-SNAPSHOT"
-java.sourceCompatibility = JavaVersion.VERSION_11
+java.sourceCompatibility = JavaVersion.VERSION_17
 
 repositories {
 	mavenCentral()
@@ -58,10 +58,11 @@ dependencies {
 
 tasks.withType<KotlinCompile> {
 	kotlinOptions {
-		freeCompilerArgs = listOf("-Xjsr305=strict")
-		jvmTarget = "11"
+		freeCompilerArgs += "-Xjsr305=strict"
+		jvmTarget = "17"
 	}
 }
+
 
 tasks.withType<Test> {
 	useJUnitPlatform()
@@ -81,7 +82,7 @@ tasks.register<JacocoReport>("jacocoRootReport") {
 	}
 
 	reports {
-		xml.outputLocation.set(File("${buildDir}/reports/jacoco/test/jacocoTestReport.xml"))
+		xml.outputLocation.set(layout.buildDirectory.file("reports/jacoco/test/jacocoTestReport.xml"))
 		xml.required.set(true)
 		html.required.set(false)
 	}
