@@ -1,6 +1,6 @@
 package team.msg.hiv2.domain.auth.domain
 
-import javax.persistence.Id
+import jakarta.persistence.Id
 import team.msg.hiv2.global.annotation.Aggregate
 import java.util.UUID
 
