@@ -1,7 +1,7 @@
 package team.msg.hiv2.domain.homebase.presentation.data.web
 
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.Size
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Size
 import java.util.*
 
 data class ReservationHomeBaseWebRequest(
