@@ -1,6 +1,6 @@
 package team.msg.hiv2.domain.reservation.presentation.data.web
 
-import javax.validation.constraints.NotNull
+import jakarta.validation.constraints.NotNull
 
 data class UpdateReservationCheckStatusWebRequest(
     @field:NotNull
