@@ -1,6 +1,6 @@
 package team.msg.hiv2.domain.user.presentation.data.request
 
-import javax.validation.constraints.NotNull
+import jakarta.validation.constraints.NotNull
 import team.msg.hiv2.domain.user.domain.constant.UserRole
 
 data class UpdateUserRoleWebRequest(
