@@ -1,8 +1,8 @@
 package team.msg.hiv2.global.entity
 
 import java.time.LocalDateTime
-import javax.persistence.Column
-import javax.persistence.MappedSuperclass
+import jakarta.persistence.Column
+import jakarta.persistence.MappedSuperclass
 
 @MappedSuperclass
 abstract class BaseTimeEntity(
