@@ -3,9 +3,9 @@ package team.msg.hiv2.global.filter
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
 import team.msg.hiv2.global.security.spi.JwtParserPort
-import javax.servlet.FilterChain
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.FilterChain
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 class JwtRequestFilter(
     private val jwtParserPort: JwtParserPort
