@@ -4,8 +4,8 @@ import org.slf4j.LoggerFactory
 import org.springframework.security.core.AuthenticationException
 import org.springframework.security.web.AuthenticationEntryPoint
 import team.msg.hiv2.global.error.exception.ForbiddenException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 class CustomAuthenticationEntryPoint : AuthenticationEntryPoint {
 
