@@ -13,7 +13,7 @@ object DependencyVersions {
     const val GAUTH_VERSION = "v3.0.0"
 
     // test
-    const val MOCKITO_VERSION = "4.0.0"
+    const val MOCKITO_VERSION = "5.0.0"
 
     // cloudwatch appender
     const val CLOUDWATCH_APPENDER_VERSION = "1.6.0"
