@@ -1,9 +1,9 @@
 object PluginVersion {
-    const val SPRING_BOOT_VERSION = "2.7.3"
-    const val DEPENDENCY_MANAGER_VERSION = "1.0.13.RELEASE"
-    const val JVM_VERSION = "1.7.10"
-    const val SPRING_PLUGIN_VERSION = "1.7.10"
-    const val JPA_PLUGIN_VERSION = "1.7.10"
-    const val KAPT_VERSION = "1.7.10"
+    const val SPRING_BOOT_VERSION = "3.1.5"
+    const val DEPENDENCY_MANAGER_VERSION = "1.1.3"
+    const val JVM_VERSION = "1.9.24"
+    const val SPRING_PLUGIN_VERSION = "1.9.24"
+    const val JPA_PLUGIN_VERSION = "1.9.24"
+    const val KAPT_VERSION = "1.9.24"
     const val QUERY_DSL_PLUGIN_VERSION = "1.0.10"
 }
