@@ -1,18 +1,18 @@
 package team.msg.hiv2.global.entity
 
+import jakarta.persistence.*
 import org.hibernate.annotations.GenericGenerator
-import java.util.UUID
-import javax.persistence.Column
-import javax.persistence.GeneratedValue
-import javax.persistence.Id
-import javax.persistence.MappedSuperclass
+import org.hibernate.annotations.JdbcTypeCode
+import org.hibernate.type.SqlTypes
+import java.util.*
 
 @MappedSuperclass
 abstract class BaseUuidEntity(
 
     @Id
     @GeneratedValue(generator = "uuid2")
-    @GenericGenerator(name = "uuid2", strategy = "uuid2")
     @Column(columnDefinition = "BINARY(16)", nullable = false)
+    @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
+    @JdbcTypeCode(SqlTypes.VARCHAR)
     open val id: UUID
 ) : BaseTimeEntity()
